requirements.txt
# Core Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pyarrow>=14.0.0  # For parquet files (ORBITAAL)

# Network Analysis
networkx>=3.1
python-louvain>=0.16  # Community detection
powerlaw>=1.5  # Power-law fitting

# Epidemic Modeling
ndlib>=5.1.0  # Network Diffusion Library
EoN>=1.1  # Epidemics on Networks

# Data Acquisition
requests>=2.28.0
pycoingecko>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning (for validation)
scikit-learn>=1.2.0

# Utilities
tqdm>=4.65.0  # Progress bars
pyyaml>=6.0  # Config files
python-dateutil>=2.8.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
